+++ packages/types/src/index.ts
@@ -4,7 +4,9 @@
  * Zero internal dependencies; everything the engine knows about "types".
  */
 export * from "./artifacts.js";
+export * from "./events.js";
 export * from "./schema.js";
 export * from "./storage.js";
+export * from "./tools.js";
 export * from "./types.js";
 export * from "./util.js";